Code refactor, move GlideModule to helper folder.
