+++ app/src/main/java/com/example/hepiplant/SalesOfferActivity.java
@@ -332,4 +332,5 @@ public class SalesOfferActivity extends AppCompatActivity implements CommentsRec
                 .load(storageRef)
                 .into(photoImageView);
     }
+
 }
